Edit chat service

+++ social_media/src/main/java/com/example/social_media/user/UserService.java
@@ -23,7 +23,7 @@ public class UserService {
         return userRepository.findAll();
     }
 
-    public User findById(Integer id) {
+    public User findById(Long id) {
         return userRepository.findById(id)
                 .orElseThrow(() -> new UserNotFoundException("User not found with id: " + id));
     }
@@ -43,7 +43,7 @@ public class UserService {
         return userRepository.save(user);
     }
 
-    public void deleteById(Integer id) {
+    public void deleteById(Long id) {
         userRepository.deleteById(id);
     }
 }
